requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
geopy
pydantic
python-multipart
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 